requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
boto3
scikit-learn
mlflow<3
pyaml
dagshub
fastapi
uvicorn
python-multipart
# -e .